+++ modules/card-project/card-project.php
@@ -10,7 +10,7 @@
 		'tag' => 'div',
 		'permalink' => bp_get_group_permalink(),
 		'cover_image_url' => $cover_image_url,
-		'tela' => false, // TODO s'il s'agit d'un projet Tela
+		'tela' => bp_groups_has_group_type( bp_get_group_id(), 'tela-botanica' ),
 		'tela_title' => __( 'Un projet Tela Botanica', 'telabotanica' ),
 		'tela_icon' => get_telabotanica_module('icon', ['icon' => 'tela-leaf']),
 		'avatar' => bp_core_fetch_avatar( [